log-analysis/requirements.txt
psycopg2-binary==2.9.9
faker==22.2.0
